+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- markdownlint-disable no-duplicate-heading -->
+## Unreleased
+
+### Breaking Changes
+
+- `HangeulInformal` has been renamed to `HanjaInformal`, its correct name.
 
 ## v0.2.2 (2023-01-18)